+++ aiter/test_mha_common.py
@@ -614,3 +614,57 @@ def attention_ref(
         attention.to(dtype=dtype_og),
         lse.to(dtype=dtype_og),
     )
+
+
+def attention_ref_with_tol(q, k, v, do, is_fp8=False, **kwargs):
+    """Run attention reference and compute adaptive tolerances.
+
+    Follows the upstream flash attention tolerance pattern
+    (see tests/test_flash_attn.py in Dao-AILab/flash-attention). Runs two
+    PyTorch references (upcast and non-upcast) and uses the gap between them as
+    a baseline for tolerance.
+
+    Returns (out, (dq, dk, dv), fwd_tol, [dq_tol, dk_tol, dv_tol])
+    where each tol is (atol, rtol).
+    """
+    has_dropout = kwargs.get("dropout_p", 0.0) > 0.0
+
+    def _run_ref(upcast, reorder_ops=False):
+        q_ = q.detach().clone().requires_grad_(True)
+        k_ = k.detach().clone().requires_grad_(True)
+        v_ = v.detach().clone().requires_grad_(True)
+        with torch.enable_grad():
+            out, _, _ = attention_ref(
+                q_, k_, v_, upcast=upcast, reorder_ops=reorder_ops, **kwargs
+            )
+        dq, dk, dv = torch.autograd.grad(out, (q_, k_, v_), do)
+        return out, dq, dk, dv
+
+    def _tol(ref_val, pt_val, is_forward=False):
+        baseline = (pt_val - ref_val).abs().max().item()
+        if is_fp8:
+            mult = 4
+            atol_floor = 5e-1 if is_forward else 1.0
+            rtol_floor = 1e-1
+        elif has_dropout:
+            # Dropout scaling (1/(1-p)) amplifies precision errors in the
+            # fused kernel differently than in the reference. The baseline
+            # between two references uses the same mask so it underestimates
+            # the kernel-vs-reference gap.
+            mult = 2
+            atol_floor = 1e-1 if is_forward else 2.0
+            rtol_floor = 1e-1
+        else:
+            mult = 2
+            atol_floor = 1e-2 if is_forward else 1.5e-2
+            rtol_floor = 1e-5
+        atol = max(mult * baseline, atol_floor)
+        return atol, rtol_floor
+
+    out, dq, dk, dv = _run_ref(upcast=True)
+    out_pt, dq_pt, dk_pt, dv_pt = _run_ref(upcast=False, reorder_ops=True)
+
+    fwd_tol = _tol(out, out_pt, is_forward=True)
+    bwd_tols = [_tol(dq, dq_pt), _tol(dk, dk_pt), _tol(dv, dv_pt)]
+
+    return out, (dq, dk, dv), fwd_tol, bwd_tols
